+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.3.0
 
-* Add `SplitGen` and `splitGen`
+* Add `Seed`, `SeedGen`, `seedSize`, `mkSeed` and `unSeed`:
+  [#162](https://github.com/haskell/random/pull/162)
+* Add `SplitGen` and `splitGen`: [#160](https://github.com/haskell/random/pull/160)
 * Add `shuffleList` and `shuffleListM`: [#140](https://github.com/haskell/random/pull/140)
 * Add `mkStdGen64`: [#155](https://github.com/haskell/random/pull/155)
 * Add `uniformListRM`, `uniformList`, `uniformListR`, `uniforms` and `uniformRs`: